limits changes for n1

+++ cloud_SQL.py
@@ -829,7 +829,7 @@ def main(sheet,email):
         
             
         if int(row["Avg no. of hrs"]) < 730:
-            one_year_price, one_year_current_url = sud_pricing(driver, actions, row["SQL Type"], row["Datacenter Location"], row["Cloud SQL "],float(row["No. of Instances"]), int(row["Avg no. of hrs"]), str(row["Instance Type"]),row["HA/Non-HA"], row["Disk Type"], int(row["Storage Amt"]),int(row["Backup"]),int(row["vCPUs"]),int(row["RAM"]))
+            one_year_price, one_year_current_url = sud_pricing(driver, actions, row["SQL Type"], row["Datacenter Location"], row["Cloud SQL "],float(row["No. of Instances"]), int(row["Avg no. of hrs"]), str(row["Instance Type"]),row["HA/Non-HA"], row["Disk Type"], int(row["Storage Amt"]),int(row["Backup"]),float(row["vCPUs"]),float(row["RAM"]))
             results[index]["One Year Price"] = one_year_price
             results[index]["One Year URL"] = one_year_current_url
             
@@ -837,7 +837,7 @@ def main(sheet,email):
             
             
         else:
-            one_year_price, one_year_current_url = one_year_pricing(driver, actions, row["SQL Type"], row["Datacenter Location"], row["Cloud SQL "],float(row["No. of Instances"]), int(row["Avg no. of hrs"]), str(row["Instance Type"]),row["HA/Non-HA"], row["Disk Type"], int(row["Storage Amt"]),int(row["Backup"]),int(row["vCPUs"]),int(row["RAM"]))
+            one_year_price, one_year_current_url = one_year_pricing(driver, actions, row["SQL Type"], row["Datacenter Location"], row["Cloud SQL "],float(row["No. of Instances"]), int(row["Avg no. of hrs"]), str(row["Instance Type"]),row["HA/Non-HA"], row["Disk Type"], int(row["Storage Amt"]),int(row["Backup"]),float(row["vCPUs"]),float(row["RAM"]))
             results[index]["One Year Price"] = one_year_price
             results[index]["One Year URL"] = one_year_current_url
         
@@ -860,7 +860,7 @@ def main(sheet,email):
     for index, row in df.iterrows():
         
         if int(row["Avg no. of hrs"]) < 730:
-            three_year_price, three_year_current_url = sud_pricing(driver, actions, row["SQL Type"], row["Datacenter Location"], row["Cloud SQL "],float(row["No. of Instances"]), int(row["Avg no. of hrs"]), str(row["Instance Type"]),row["HA/Non-HA"], row["Disk Type"], int(row["Storage Amt"]),int(row["Backup"]),int(row["vCPUs"]),int(row["RAM"]))
+            three_year_price, three_year_current_url = sud_pricing(driver, actions, row["SQL Type"], row["Datacenter Location"], row["Cloud SQL "],float(row["No. of Instances"]), int(row["Avg no. of hrs"]), str(row["Instance Type"]),row["HA/Non-HA"], row["Disk Type"], int(row["Storage Amt"]),int(row["Backup"]),float(row["vCPUs"]),float(row["RAM"]))
             results[index]["three Year Price"] = three_year_price
             results[index]["three Year URL"] = three_year_current_url
             
@@ -868,7 +868,7 @@ def main(sheet,email):
             
             
         else:
-            three_year_price, three_year_current_url = three_year_pricing(driver, actions, row["SQL Type"], row["Datacenter Location"], row["Cloud SQL "],float(row["No. of Instances"]), int(row["Avg no. of hrs"]), str(row["Instance Type"]),row["HA/Non-HA"], row["Disk Type"], int(row["Storage Amt"]),int(row["Backup"]),int(row["vCPUs"]),int(row["RAM"]))
+            three_year_price, three_year_current_url = three_year_pricing(driver, actions, row["SQL Type"], row["Datacenter Location"], row["Cloud SQL "],float(row["No. of Instances"]), int(row["Avg no. of hrs"]), str(row["Instance Type"]),row["HA/Non-HA"], row["Disk Type"], int(row["Storage Amt"]),int(row["Backup"]),float(row["vCPUs"]),float(row["RAM"]))
             results[index]["three Year Price"] = three_year_price
             results[index]["three Year URL"] = three_year_current_url
